Split the H/2 raw frame tests into its own file
